Fix clang warning about local types being used as template parameters.
clang would complain

/Users/thakis/src/chrome-git/src/net/base/host_resolver_impl.cc:557:5: error: template argument uses local type 'net::HostResolverImpl::Job::Category' [-Wlocal-type-template-args]
    DCHECK_LT(category, RESOLVE_MAX);  // Be sure it was set.
    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In file included from /Users/thakis/src/chrome-git/src/net/base/host_resolver_impl.cc:5:
In file included from ../net/base/host_resolver_impl.h:18:
In file included from ../net/base/network_change_notifier.h:10:
In file included from ../base/observer_list_threadsafe.h:14:
../base/logging.h:662:31: note: instantiated from:
#define DCHECK_LT(val1, val2) DCHECK_OP(LT, < , val1, val2)
                              ^
../base/logging.h:635:9: note: instantiated from:
        logging::Check##name##Impl((val1), (val2),              \
        ^~~~~~~

This seems to be a new warning. Since it only affects this file, we might as well fix the problem instead of disabling the warning.

BUG=None
TEST=clang/mac waterfall stays green after clobber build.

Review URL: http://codereview.chromium.org/3816009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@62934 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/host_resolver_impl.cc
@@ -554,7 +554,7 @@ class HostResolverImpl::Job
                                        std::abs(os_error_),
                                        GetAllGetAddrinfoOSErrors());
     }
-    DCHECK_LT(category, RESOLVE_MAX);  // Be sure it was set.
+    DCHECK_LT(category, static_cast<int>(RESOLVE_MAX));  // Be sure it was set.
 
     UMA_HISTOGRAM_ENUMERATION("DNS.ResolveCategory", category, RESOLVE_MAX);
